% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rustup.R
\name{rustup_windows}
\alias{rustup_windows}
\title{Install Rust}
\usage{
rustup_windows(targets = c("i686-pc-windows-gnu", "x86_64-pc-windows-gnu"))
}
\arguments{
\item{targets}{which compiler targets would you like to install}
}
\description{
Install Rust
}
\examples{
rustup_windows()
}
